0.9.2 release.
Improve look of the title bar on Linux, Mac OS and Windows. The bar acts and looks a lot closer to a native window on the respective operating system. However this is limited to dark and light modes of one particular style. It cannot determine operating system configuration of the theme beyond dark / light mode (and even this not on Linux).

 * Windows mimics Windows 11
 * Mac OS mimics Yosemite
 * Linux mimics Yaru (Ubuntu 23.10)